# server/docs/swagger.yaml
definitions:
...
  datatypes.JSONMap:
    additionalProperties: true
    type: object
...
  github_com_test-instructor_yangfan_server_model_interfacecase.ApiCase:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      RunConfigID:
        type: integer
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      api_env_id:
        type: integer
      api_env_name:
        type: string
      case:
        items:
          $ref: '#/definitions/interfacecase.ApiCaseStep'
        type: array
      describe:
        type: string
      front_case:
        description: Performance []Performance   `json:"performance,omitempty" form:"performance"
          gorm:"many2many:PerformanceRelationship;"`
        type: boolean
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      project_id:
        type: integer
      runConfig:
        $ref: '#/definitions/interfacecase.ApiConfig'
      runNumber:
        type: integer
      status:
        type: boolean
    type: object
  interfacecase.ApiCaseStep:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      RunConfigID:
        type: integer
      RunConfigName:
        type: string
      TStep:
        items:
          $ref: '#/definitions/interfacecase.ApiStep'
        type: array
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      api_env_id:
        type: integer
      api_env_name:
        type: string
      api_step_type:
        $ref: '#/definitions/interfacecase.ApiStepType'
      case:
        items:
          $ref: '#/definitions/github_com_test-instructor_yangfan_server_model_interfacecase.ApiCase'
        type: array
      front_case:
        type: boolean
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      project_id:
        type: integer
      type:
        $ref: '#/definitions/interfacecase.ApiType'
    type: object
  interfacecase.ApiConfig:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      allow_redirects:
        type: boolean
      base_url:
        type: string
      case_id:
        type: integer
      default:
        type: boolean
      environs:
        additionalProperties:
          type: string
        description: environment variables
        type: object
      export:
        items:
          type: string
        type: array
      headers:
        $ref: '#/definitions/datatypes.JSONMap'
      headers_json:
        items:
          type: integer
        type: array
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      parameters:
        $ref: '#/definitions/datatypes.JSONMap'
      project_id:
        type: integer
      report_id:
        type: integer
      setup_case:
        $ref: '#/definitions/interfacecase.ApiCaseStep'
      setup_case_id:
        type: integer
      timeout:
        type: number
      variables:
        $ref: '#/definitions/datatypes.JSONMap'
      variables_json:
        items:
          type: integer
        type: array
      verify:
        type: boolean
      weight:
        type: integer
    type: object
  interfacecase.ApiGrpc:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      body:
        $ref: '#/definitions/datatypes.JSONMap'
      detail:
        $ref: '#/definitions/datatypes.JSONMap'
      headers:
        $ref: '#/definitions/datatypes.JSONMap'
      headers_json:
        items:
          type: integer
        type: array
      id:
        description: 主键ID
        type: integer
      timeout:
        type: number
      type:
        $ref: '#/definitions/interfacecase.GrpcType'
      url:
        type: string
    type: object
...
  interfacecase.ApiRequest:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      agreement:
        type: string
      allow_redirects:
        type: boolean
      data:
        $ref: '#/definitions/datatypes.JSONMap'
      data_json:
        items:
          type: integer
        type: array
      headers:
        $ref: '#/definitions/datatypes.JSONMap'
      headers_json:
        items:
          type: integer
        type: array
      http2:
        type: boolean
      id:
        description: 主键ID
        type: integer
      json:
        items:
          type: integer
        type: array
      method:
        type: string
      params:
        $ref: '#/definitions/datatypes.JSONMap'
      params_json:
        items:
          type: integer
        type: array
      timeout:
        type: number
      url:
        type: string
      verify:
        type: boolean
    type: object
  interfacecase.ApiStep:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      export_header:
        items:
          type: integer
        type: array
      export_parameter:
        items:
          type: integer
        type: array
      extract:
        $ref: '#/definitions/datatypes.JSONMap'
      extract_json:
        items:
          type: integer
        type: array
      gRPC:
        $ref: '#/definitions/interfacecase.ApiGrpc'
      grpc_id:
        type: integer
      hooks:
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      project_id:
        type: integer
      rendezvous:
        $ref: '#/definitions/interfacecase.ApiStepRendezvous'
      rendezvous_id:
        type: integer
      request:
        $ref: '#/definitions/interfacecase.ApiRequest'
      request_id:
        type: integer
      setup_hooks:
        items:
          type: string
        type: array
      sort:
        type: integer
      teardown_hooks:
        items:
          type: string
        type: array
      testCase:
        items:
          $ref: '#/definitions/interfacecase.ApiCaseStep'
        type: array
      think_time:
        $ref: '#/definitions/interfacecase.ApiStepThinkTime'
      think_time_id:
        type: integer
      transaction:
        $ref: '#/definitions/interfacecase.ApiStepTransaction'
      transaction_id:
        type: integer
      type:
        $ref: '#/definitions/interfacecase.ApiType'
      validate:
        items:
          additionalProperties: true
          type: object
        type: array
      validate_json:
        items:
          type: integer
        type: array
      validate_number:
        type: integer
      variables:
        $ref: '#/definitions/datatypes.JSONMap'
      variables_json:
        items:
          type: integer
        type: array
    type: object
  interfacecase.ApiStepRendezvous:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: required
        type: string
      number:
        type: integer
      percent:
        description: default to 1(100%)
        type: number
      timeout:
        description: milliseconds
        type: integer
    type: object
  interfacecase.ApiStepThinkTime:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      id:
        description: 主键ID
        type: integer
      time:
        type: number
    type: object
  interfacecase.ApiStepTransaction:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/interfacecase.TransactionType'
    type: object
  interfacecase.ApiStepType:
    enum:
    - Transaction
    - TransactionStart
    - TransactionEnd
    - Rendezvous
    type: string
    x-enum-varnames:
    - ApiStepTypeTransaction
    - ApiStepTypeTransactionStart
    - ApiStepTypeTransactionEnd
    - ApiStepTypeRendezvous
  interfacecase.ApiTimerTask:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      RunConfigID:
        type: integer
      TestCase:
        items:
          $ref: '#/definitions/github_com_test-instructor_yangfan_server_model_interfacecase.ApiCase'
        type: array
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      api_env_id:
        type: integer
      api_env_name:
        type: string
      api_message_id:
        type: integer
      apiTimerTaskTag:
        items:
          $ref: '#/definitions/interfacecase.ApiTimerTaskTag'
        type: array
      config:
        $ref: '#/definitions/interfacecase.ApiConfig'
      describe:
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      nextRunTime:
        type: string
      project_id:
        type: integer
      runNumber:
        type: integer
      runTime:
        type: string
      status:
        type: boolean
      tagIds:
        items:
          type: integer
        type: array
    type: object
  interfacecase.ApiTimerTaskTag:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      CreatedBy:
        type: integer
      DeleteBy:
        type: integer
      UpdateBy:
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      apiTimerTask:
        items:
          $ref: '#/definitions/interfacecase.ApiTimerTask'
        type: array
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      project_id:
        type: integer
      remarks:
        type: string
    type: object
  interfacecase.ApiType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      ApiTypeCase: 接口测试用例
      ApiTypeCasePerformance: 性能测试用例
      ApiTypeTemplate: 接口模板
    x-enum-varnames:
    - ApiTypeTemplate
    - ApiTypeCase
    - ApiTypeCasePerformance
  interfacecase.GrpcType:
    enum:
    - Simple
    - ServerSideStream
    - ClientSideStream
    - BidirectionalStream
    type: string
    x-enum-varnames:
    - GrpcTypeSimple
    - GrpcTypeServerSideStream
    - GrpcTypeClientSideStream
    - GrpcTypeBidirectionalStream
  interfacecase.TransactionType:
    enum:
    - start
    - end
    type: string
    x-enum-varnames:
    - TransactionStart
    - TransactionEnd
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /task/createTimerTask:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建TimerTask
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/interfacecase.ApiTimerTask'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建TimerTask
      tags:
      - TimerTask
  /task/deleteTimerTask:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除TimerTask
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/interfacecase.ApiTimerTask'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除TimerTask
      tags:
      - TimerTask
  /task/deleteTimerTaskByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除TimerTask
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除TimerTask
      tags:
      - TimerTask
  /task/findTimerTask:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - in: query
        name: CreatedBy
        type: integer
      - in: query
        name: DeleteBy
        type: integer
      - in: query
        name: RunConfigID
        type: integer
      - in: query
        name: UpdateBy
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: api_env_id
        type: integer
      - in: query
        name: api_env_name
        type: string
      - in: query
        name: api_message_id
        type: integer
      - in: query
        name: describe
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: nextRunTime
        type: string
      - in: query
        name: project_id
        type: integer
      - in: query
        name: runNumber
        type: integer
      - in: query
        name: runTime
        type: string
      - in: query
        name: status
        type: boolean
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: tagIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询TimerTask
      tags:
      - TimerTask
  /task/getTimerTaskList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - in: query
        name: CreatedBy
        type: integer
      - in: query
        name: DeleteBy
        type: integer
      - in: query
        name: RunConfigID
        type: integer
      - in: query
        name: UpdateBy
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: api_env_id
        type: integer
      - in: query
        name: api_env_name
        type: string
      - in: query
        name: api_message_id
        type: integer
      - in: query
        name: describe
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: nextRunTime
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: project_id
        type: integer
      - in: query
        name: runNumber
        type: integer
      - in: query
        name: runTime
        type: string
      - in: query
        name: status
        type: boolean
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: tagIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取TimerTask列表
      tags:
      - TimerTask
  /task/updateTimerTask:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新TimerTask
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/interfacecase.ApiTimerTask'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新TimerTask
      tags:
      - TimerTask
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"